+++ Mirror/Runtime/NetworkServer.cs
@@ -85,17 +85,17 @@ namespace Mirror
             RegisterHandler(MsgType.Ping, NetworkTime.OnServerPing);
         }
 
-        public static bool Listen(int serverPort, int maxConnections)
+        public static bool Listen(int serverPort, ushort maxConnections)
         {
             return InternalListen(null, serverPort, maxConnections);
         }
 
-        public static bool Listen(string ipAddress, int serverPort, int maxConnections)
+        public static bool Listen(string ipAddress, int serverPort, ushort maxConnections)
         {
             return InternalListen(ipAddress, serverPort, maxConnections);
         }
 
-        internal static bool InternalListen(string ipAddress, int serverPort, int maxConnections)
+        internal static bool InternalListen(string ipAddress, int serverPort, ushort maxConnections)
         {
             Initialize();
